refactor: isolate accumulator limit accounting
